API

RandomisationModels Module

RandomisationModels.MassWeightedUrn — Type

MassWeightedUrn <: MultiArmRandomisationModel

A mass-weighted urn randomisation model.

RandomisationModels.MassWeightedUrn — Method

MassWeightedUrn(target::AbstractVector{T}, α::Real) where {T<:Real}

RandomisationModels.randomise! — Method

randomise!(rng::AbstractRNG, RM::RandomisationModel, n::Int)

Generate a n random allocation's from a randomisation model using prob(RM) and the random number generator provided.

RandomisationModels.randomise! — Method

randomise!(rng::AbstractRNG, RM::RandomisationModel)

Generate a single random allocation from a randomisation model, RM, using prob(RM) and the random number generator provided.

RandomisationModels.randomise! — Method

randomise!(RM::RandomisationModel, n::Int)

Generate a n random allocation's from a randomisation model using prob(RM) and the global RNG.

RandomisationModels.randomise! — Method

randomise!(RM::RandomisationModel)

Generate a single random allocation from a randomisation model using prob(RM) and the global RNG.

RandomisationModels.randomise — Method

randomise(rng::AbstractRNG, RM::RandomisationModel, n::Int)

Generates n allocations

RandomisationModels.randomise — Method

randomise(rng::AbstractRNG, RM::RandomisationModel)

Generates allocations from the randomisation model without updating the models.

RandomisationModels.randomize — Method

randomize(rng::AbstractRNG, RM::RandomisationModel)

Generates allocations from the randomisation model without updating the models.

Util Sub-Module

RandomisationModels.Util — Module

Util

A submodule that provides various utility functions used elsewhere in the main Module.

RandomisationModels.Util.convert_prob_to_intweight — Method

convert_prob_to_intweight(target::Vector{<:Real})

Convert a vector to a vector of integer weights

RandomisationModels.Util.isnormvec — Method

isnormvec(p)

Check whether in abstract vector p sums to 1.

RandomisationModels.Util.isposvec — Method

isposvec(p)

Check whether in abstract vector p is a positive vector, i.e. that pᵢ≥0∀i.

RandomisationModels.Util.isprobvec — Method

isprobvec(p)

Check whether in abstract vector p is a probability vector, i.e. that ∑p=1 and pᵢ≥0∀i.