API

RandomisationModels.BlockUrn — Type

struct BlockUrn <: MultiArmRandomisationModel

Implements the Block Urn Design (BUD) of Zhao and Wang (2011).

The BUD has conditional allocation probability

\[P_{im} = \frac{w_m\lambda + w_mk_{i-1}-n_{i-1,m}}{W\lambda + Wk_{i-1} - (i - 1)}, \quad k_{i-1} = \min_{m=1,...,M} \left\lfloor \frac{n_{i-1,k}}{w_k} \right\rfloor.\]

Fields

target::AbstractVector{<:AbstractFloat}

blocksize::Integer

weights::AbstractVector{<:Integer}

W::Integer

sequence::AbstractVector{<:Integer}

dist::AbstractVector{<:Integer}

RandomisationModels.BlockUrn — Method

BlockUrn(
    target::AbstractVector{<:AbstractFloat},
    blocksize::Int64
) -> BlockUrn

RandomisationModels.BlockUrn — Method

BlockUrn(
    weights::AbstractVector{<:Integer},
    blocksize::Int64
) -> BlockUrn

RandomisationModels.CompleteRandomisation — Type

struct CompleteRandomisation{Ts<:(AbstractVector)} <: MultiArmRandomisationModel

Fields

target::AbstractVector

sequence::AbstractVector{Int64}

dist::AbstractVector{Int64}

RandomisationModels.CompleteRandomisation — Method

CompleteRandomisation(
    target::AbstractArray{T<:Real, 1}
) -> CompleteRandomisation

RandomisationModels.MassWeightedUrn — Type

struct MassWeightedUrn{Tv<:(AbstractVector), T<:Real} <: MultiArmRandomisationModel

Implements the Mass Weighted Urn Design (MWUD) of Zhao (2015).

The MWU has conditional allocation probability

\[P_{im} = \frac{\max{\alpha w_m - n_{i-1,m} + (i - 1)w_m, 0}}{\sum \max{\alpha w_m - n_{i-1,m} + (i - 1)w_m, 0}}\]

Fields

target::AbstractVector

α::Real

mass::AbstractVector

sequence::AbstractVector{Int64}

dist::AbstractVector{Int64}

RandomisationModels.MassWeightedUrn — Method

RandomisationModels.MassWeightedUrn — Method

MassWeightedUrn(
    target::AbstractArray{T<:Real, 1},
    α::Real
) -> MassWeightedUrn

RandomisationModels.PermutedBlock — Type

struct PermutedBlock <: MultiArmRandomisationModel

Fields

target::AbstractVector{<:AbstractFloat}

blocksize::Integer

weights::AbstractVector{<:Integer}

W::Integer

sequence::AbstractVector{<:Integer}

dist::AbstractVector{<:Integer}

RandomisationModels.PermutedBlock — Method

PermutedBlock(
    target::AbstractVector{<:AbstractFloat},
    blocksize::Int64
) -> PermutedBlock

RandomisationModels.PermutedBlock — Method

PermutedBlock(
    weights::AbstractVector{<:Integer},
    blocksize::Int64
) -> PermutedBlock

RandomisationModels.prob — Method

prob(BU::BlockUrn) -> Any

RandomisationModels.prob — Method

prob(PB::PermutedBlock) -> Any

RandomisationModels.randomise! — Method

randomise!(rng::AbstractRNG, RM::RandomisationModel, n::Int)

Generate a n random allocation's from a randomisation model using prob(RM) and the random number generator provided.

RandomisationModels.randomise! — Method

randomise!(rng::AbstractRNG, RM::RandomisationModel)

Generate a single random allocation from a randomisation model, RM, using prob(RM) and the random number generator provided.

RandomisationModels.randomise! — Method

randomise!(RM::RandomisationModel, n::Int)

Generate a n random allocation's from a randomisation model using prob(RM) and the global RNG.

RandomisationModels.randomise! — Method

randomise!(RM::RandomisationModel)

Generate a single random allocation from a randomisation model using prob(RM) and the global RNG.

RandomisationModels.randomise — Method

randomise(rng::AbstractRNG, RM::RandomisationModel, n::Int)

Generates n allocations

RandomisationModels.randomise — Method

randomise(rng::AbstractRNG, RM::RandomisationModel)

Generates allocations from the randomisation model without updating the models.

RandomisationModels.randomize — Method

randomize(rng::AbstractRNG, RM::RandomisationModel)

Generates allocations from the randomisation model without updating the models.

RandomisationModels.update! — Method

update!(BU::BlockUrn, y::Int64)

RandomisationModels.update! — Method

update!(CR::CompleteRandomisation, y::Int64)

RandomisationModels.update! — Method

update!(MWU::MassWeightedUrn, y::Int64)

RandomisationModels.update! — Method

update!(PB::PermutedBlock, y::Int64)